+++ src/vulkan/vk-device.cpp
@@ -1402,7 +1402,7 @@ Result DeviceImpl::getTextureAllocationInfo(const TextureDesc& desc_, Size* outS
     case TextureType::Texture1DArray:
     {
         imageInfo.imageType = VK_IMAGE_TYPE_1D;
-        imageInfo.extent = VkExtent3D{uint32_t(desc.size.width), 1, 1};
+        imageInfo.extent = VkExtent3D{desc.size.width, 1, 1};
         break;
     }
     case TextureType::Texture2D:
@@ -1411,7 +1411,7 @@ Result DeviceImpl::getTextureAllocationInfo(const TextureDesc& desc_, Size* outS
     case TextureType::Texture2DMSArray:
     {
         imageInfo.imageType = VK_IMAGE_TYPE_2D;
-        imageInfo.extent = VkExtent3D{uint32_t(desc.size.width), uint32_t(desc.size.height), 1};
+        imageInfo.extent = VkExtent3D{desc.size.width, desc.size.height, 1};
         break;
     }
     case TextureType::Texture3D:
@@ -1419,14 +1419,14 @@ Result DeviceImpl::getTextureAllocationInfo(const TextureDesc& desc_, Size* outS
         // Can't have an array and 3d texture
         SLANG_RHI_ASSERT(desc.arrayLength <= 1);
         imageInfo.imageType = VK_IMAGE_TYPE_3D;
-        imageInfo.extent = VkExtent3D{uint32_t(desc.size.width), uint32_t(desc.size.height), uint32_t(desc.size.depth)};
+        imageInfo.extent = VkExtent3D{desc.size.width, desc.size.height, desc.size.depth};
         break;
     }
     case TextureType::TextureCube:
     case TextureType::TextureCubeArray:
     {
         imageInfo.imageType = VK_IMAGE_TYPE_2D;
-        imageInfo.extent = VkExtent3D{uint32_t(desc.size.width), uint32_t(desc.size.height), 1};
+        imageInfo.extent = VkExtent3D{desc.size.width, desc.size.height, 1};
         imageInfo.flags = VK_IMAGE_CREATE_CUBE_COMPATIBLE_BIT;
         break;
     }
